Stop Rubocop looking for configs & violations in vendor/ folder
By default, this folder is excluded [issue], however when we override
the `AllCops.Exclude` option it is discarded. On CI, where we put our
gems in vendor/bundle, Rubocop was looking at all ~4492 files in those
directories and linting them. It was also parsing the .rubocop.yml files
in those directories and trying to use that configuration, leading to
this mysterious error:

  Running RuboCop...
  cannot load such file -- rubocop-rspec

(even though we do not use this library in our gem!)

By setting `inherit_mode` in .rubocop.yml, we tell it to use the default
set of exclusions, defined here: [global].

[issue]: rubocop/rubocop#288 (comment)
[global]: https://github.com/rubocop/rubocopblob/930c812d3706a2cd647bdb20cb9c4a6d7a7ae35d/config/default.yml#L63-L67/